requirements.txt
python-dotenv
psycopg2-binary
git+https://github.com/pdftables/python-pdftables-api.git
pytest